[*] rename global.h to data_game.h
